Remove bash version
bash is not suited for this.  that version has been removed.

Improve the Ctrl+C detection to exit faster and add messages to indicate the progress of closing jobs/processes.
Store the estimated time left once it stabilizes and display it along with the actual script execution time.
Change the color of the message indicating where the block sizes have been written to code 121 for 256 color codes.
Change the color of the console output title to color code 74 for 256 color codes.